+++ plugins/scaffolder-backend-module-confluence-to-markdown/src/module.ts
@@ -17,7 +17,7 @@ import {
   createBackendModule,
   coreServices,
 } from '@backstage/backend-plugin-api';
-import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node/alpha';
+import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node';
 import { createConfluenceToMarkdownAction } from './actions';
 import { ScmIntegrations } from '@backstage/integration';
