+++ plugins/scaffolder-backend-module-cookiecutter/src/module.ts
@@ -17,7 +17,7 @@ import {
   coreServices,
   createBackendModule,
 } from '@backstage/backend-plugin-api';
-import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node/alpha';
+import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node';
 import { createFetchCookiecutterAction } from './actions';
 import { ScmIntegrations } from '@backstage/integration';
