+++ plugins/scaffolder-backend-module-yeoman/src/module.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import { createBackendModule } from '@backstage/backend-plugin-api';
-import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node/alpha';
+import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node';
 import { createRunYeomanAction } from './actions';
 
 /**
