+++ plugins/scaffolder-node/src/alpha/index.ts
@@ -17,7 +17,6 @@
 import { createExtensionPoint } from '@backstage/backend-plugin-api';
 import {
   TaskBroker,
-  TemplateAction,
   TemplateFilter,
   TemplateGlobal,
 } from '@backstage/plugin-scaffolder-node';
@@ -29,25 +28,6 @@ export * from './filters';
 export * from './globals';
 export * from './types';
 export * from './checkpoints';
-
-/**
- * Extension point for managing scaffolder actions.
- *
- * @alpha
- */
-export interface ScaffolderActionsExtensionPoint {
-  addActions(...actions: TemplateAction<any, any, any>[]): void;
-}
-
-/**
- * Extension point for managing scaffolder actions.
- *
- * @alpha
- */
-export const scaffolderActionsExtensionPoint =
-  createExtensionPoint<ScaffolderActionsExtensionPoint>({
-    id: 'scaffolder.actions',
-  });
 
 /**
  * Extension point for replacing the scaffolder task broker.
@@ -63,6 +43,7 @@ export interface ScaffolderTaskBrokerExtensionPoint {
  * Extension point for replacing the scaffolder task broker.
  *
  * @alpha
+ * @deprecated this extension point is planned to be removed, please reach out to us in an issue if you're using this extension point and your use cases.
  */
 export const scaffolderTaskBrokerExtensionPoint =
   createExtensionPoint<ScaffolderTaskBrokerExtensionPoint>({
